functions/fetcher/requirements.txt
google-cloud-bigquery
requests
cloudevents 